Bluepill w/ Xcode 11 support, more logging, minor fixes and partial revert of bugfix in PR#338 (#361)
* Add more logging

* Minor fixes to fix some warnings and fixing the bluepill.sh for missing xcpretty tool

* Partial revert of "Fix wrong exit status (linkedin#338)"

+++ bp/src/main.m
@@ -66,10 +66,10 @@ int main(int argc, char * argv[]) {
             NSString *fileName = [NSString stringWithFormat:@"%@-stats.json", [[config.testBundlePath lastPathComponent] stringByDeletingPathExtension]];
             NSString *outputFile = [config.outputDirectory stringByAppendingPathComponent:fileName];
             BPWriter *statsWriter = [[BPWriter alloc] initWithDestination:BPWriterDestinationFile andPath:outputFile];
-            [[BPStats sharedStats] exitWithWriter:statsWriter exitCode:exitCode];
+            [[BPStats sharedStats] exitWithWriter:statsWriter exitCode:(int)exitCode];
         }
 
         [BPUtils printInfo:INFO withString:@"BP exiting %ld", (long)exitCode];
-        return exitCode;
+        return (int)exitCode;
     }
 }
